Set up form to load and send dev profile data to backend
