+++ backend/app/controllers/users_controller.rb
@@ -21,7 +21,7 @@ class UsersController < ApplicationController
     json_response(@user, :created)
   end
 
-  # PUT /users/:id
+  # PATCH or PUT /users/:id
   def update
     @user.update!(user_params)
     head :no_content
@@ -43,7 +43,7 @@ class UsersController < ApplicationController
 
   def user_params
     # whitelist params
-    params.permit(:name, :img, :is_dev, :dev_bio, :dev_twitter, :dev_github, :dev_linkedin, :dev_portfolio)
+    params.permit(:id, :name, :img, :is_dev, :dev_bio, :dev_twitter, :dev_github, :dev_linkedin, :dev_portfolio)
   end
 
   def set_user
